+++ src/app/(pages)/sign-in/page.tsx
@@ -1,7 +1,1 @@
-"use client";
-import { useTranslation } from "react-i18next";
-
-export default function SignIn() {
-  const { t } = useTranslation();
-  return <div>{t("sign_in.text")}</div>;
-}
+export { default } from "@/app/components/sign-in/sign-in";
