+++ src/app/(pages)/sign-up/page.tsx
@@ -1,3 +1,1 @@
-export default function SignUp() {
-  return <div>sign up</div>;
-}
+export { default } from "@/app/components/sign-up/sign-up";
