+++ src/app/components/sign-in/sign-in.tsx
@@ -1,0 +1,7 @@
+"use client";
+import { useTranslation } from "react-i18next";
+
+export default function SignIn() {
+  const { t } = useTranslation();
+  return <div>{t("sign_in.text")}</div>;
+}
