+++ src/app/components/sign-up/sign-up.tsx
@@ -1,0 +1,3 @@
+export default function SignUp() {
+  return <div>sign up</div>;
+}
