+++ src/app/internationalization/config.ts
@@ -4,7 +4,7 @@ import i18n from "i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
 import Backend from "i18next-http-backend";
 import { initReactI18next } from "react-i18next";
-import { ProcessEnv } from "../types/enums";
+import { ProcessEnv } from "@/app/types/enums";
 
 export const initI18n = async () => {
   await i18n
